--- server/utils/enums.js
@@ -1,7 +1,0 @@
-export const TICKET_STATUS = {
-  NEW: 'New',
-  RESOLVED: 'Resolved',
-  CLOSED: 'Closed',
-  PENDING: 'Pending',
-  PROCESSING: 'Processing',
-};
